chore: major refactor to eliminate duplication and fix reliability
