+++ src/components/Landing/LandingTranslations.ts
@@ -1,0 +1,404 @@
+export const LANDING_COPY: Record<string, any> = {
+  'pt-BR': {
+    nav: {
+      features: 'Recursos',
+      how: 'Como funciona',
+      faq: 'FAQ',
+      github: 'GitHub',
+      signin: 'Entrar',
+      cta: 'Cadastrar',
+    },
+    hero: {
+      badge: '100% open source · gratuito',
+      h1a: 'Seu dinheiro,',
+      h1b: 'sob controle.',
+      sub: 'Rastreie despesas, receitas e investimentos em um só lugar. Importe extratos do seu banco, categorize automaticamente e entenda para onde vai cada real.',
+      cta: 'Começar grátis',
+      cta2: 'Ver demo',
+      trust: 'Sem cartão de crédito · Dados criptografados · Código aberto',
+    },
+    problems: {
+      kicker: 'O problema',
+      title: 'Você já passou pelo mês sem saber pra onde foi o dinheiro?',
+      items: [
+        {
+          t: 'Planilhas que ninguém atualiza',
+          d: 'Começa com disciplina na primeira semana. Some até o fim do mês.',
+        },
+        {
+          t: '"Para onde foi meu salário?"',
+          d: 'Fatura do cartão chega e você não lembra de 80% das compras.',
+        },
+        {
+          t: 'Sem visão do todo',
+          d: 'Investimentos de um lado, gastos do outro, reservas no meio. Nada se conversa.',
+        },
+      ],
+    },
+    solution: {
+      kicker: 'A solução',
+      title: 'Um dashboard. Todo o seu dinheiro.',
+      sub: 'Expense Tracker une suas contas, cartões e categorias em uma única visão que entende tanto o seu mês quanto o seu saldo acumulado total.',
+    },
+    features: {
+      kicker: 'Recursos',
+      title: 'Tudo o que você precisa. Nada do que não vai usar.',
+      items: [
+        {
+          tag: '01',
+          t: 'Dashboard inteligente',
+          d: 'Saldo disponível e "Caixinhas" refletem o valor real acumulado, enquanto receitas e despesas respeitam o filtro mensal.',
+        },
+        {
+          tag: '02',
+          t: 'Importação bancária (CSV)',
+          d: 'Assistente de importação com detecção automática de categorias e mapeamento de colunas do extrato.',
+        },
+        {
+          tag: '03',
+          t: 'Categorias customizáveis',
+          d: 'Categorias nativas + crie as suas com emoji e cor. Personalize do seu jeito.',
+        },
+        {
+          tag: '04',
+          t: 'Dark mode nativo',
+          d: 'Tema escuro pra quem fecha as contas de madrugada. Alterna com um clique.',
+        },
+        {
+          tag: '05',
+          t: 'Segurança Supabase',
+          d: 'Autenticação completa (OAuth, recuperação), banco criptografado e RLS em todas as tabelas.',
+        },
+      ],
+    },
+    how: {
+      kicker: 'Como funciona',
+      title: 'Do cadastro ao primeiro relatório em menos de 5 minutos.',
+      steps: [
+        {
+          n: '01',
+          t: 'Crie sua conta',
+          d: 'E-mail e senha ou login com Google. Sem papelada, sem cartão.',
+        },
+        {
+          n: '02',
+          t: 'Importe suas transações',
+          d: 'Suba o CSV do seu banco. A gente detecta categorias automaticamente.',
+        },
+        {
+          n: '03',
+          t: 'Visualize seus relatórios',
+          d: 'Gráficos de receita vs. despesa, por categoria, por mês. Tudo em tempo real.',
+        },
+        {
+          n: '04',
+          t: 'Tome decisões',
+          d: 'Entenda seu padrão de gastos e reorganize o que importa.',
+        },
+      ],
+    },
+    cta: {
+      title: 'Pronto pra controlar suas finanças?',
+      sub: 'Leva 5 minutos pra começar. Você pode fechar esta aba depois.',
+      btn: 'Começar grátis',
+      btn2: 'Ver no GitHub',
+    },
+    footer: {
+      tag: 'Feito com ❤️ por Davyd Fontoura',
+      product: 'Produto',
+      company: 'Projeto',
+      resources: 'Recursos',
+      links: {
+        product: [
+          ['Recursos', '#features'],
+          ['Como funciona', '#how'],
+          ['Demo', '#'],
+          ['Dark mode', '#'],
+        ],
+        company: [
+          ['GitHub', 'https://github.com/davydfontourac/expense-tracker'],
+          ['Issues', 'https://github.com/davydfontourac/expense-tracker/issues'],
+          ['Licença MIT', '#'],
+        ],
+        resources: [
+          ['Documentação', '#'],
+          ['Changelog', '#'],
+          ['Privacidade', '#'],
+        ],
+      },
+      rights: '© 2026 Expense Tracker · Open source sob licença MIT',
+    },
+    mock: {
+      months: ['Jan', 'Fev', 'Mar', 'Abr', 'Mai', 'Jun', 'Jul', 'Ago', 'Set', 'Out', 'Nov', 'Dez'],
+      sidebar: {
+        general: 'GERAL',
+        dashboard: 'Dashboard',
+        transactions: 'Transações',
+        categories: 'Categorias',
+        savings: 'Caixinhas',
+        analysis: 'ANÁLISES',
+        reports: 'Relatórios',
+        goals: 'Metas',
+        insights: 'Insights',
+        activeFilter: 'Filtro ativo',
+        october: 'OUTUBRO 2026',
+      },
+      header: {
+        greeting: 'Boa tarde, davyd.',
+        spent: 'Você gastou',
+        ofRevenue: 'da sua receita',
+        inApril: 'em abril',
+        april2026: 'Abril 2026',
+        all: 'Todos',
+        new: 'Nova',
+      },
+      kpis: {
+        available: 'DISPONÍVEL',
+        balance: 'Saldo em conta',
+        income: 'RECEITAS',
+        thisMonth: 'Neste mês',
+        expenses: 'DESPESAS',
+        vsLastMonth: 'vs. mês anterior',
+      },
+      charts: {
+        monthlyTitle: 'Evolução Mensal',
+        monthlySub: 'Receitas vs. Despesas · 12 meses',
+        janDez: 'JAN → DEZ',
+        income: 'Receitas',
+        expenses: 'Despesas',
+        distTitle: 'Distribuição',
+        spent: 'GASTO',
+      },
+      dist: {
+        debit: 'Débito',
+        pix: 'Pix',
+        food: 'Alimentação',
+        transport: 'Transporte',
+        others: 'Outros',
+      },
+    },
+    visuals: {
+      balance: 'SALDO',
+      accumulated: 'ACUMULADO',
+      expenses: 'DESPESAS',
+      oct2026: 'OUT/2026',
+      csv: {
+        food: 'Alimentação',
+        transport: 'Transporte',
+        subs: 'Assinaturas',
+        imported: 'transações importadas',
+      },
+      categories: [
+        'Alimentação',
+        'Transporte',
+        'Moradia',
+        'Lazer',
+        'Saúde',
+        'Educação',
+        'Viagem',
+        'Nova',
+      ],
+    },
+  },
+  en: {
+    nav: {
+      features: 'Features',
+      how: 'How it works',
+      faq: 'FAQ',
+      github: 'GitHub',
+      signin: 'Sign in',
+      cta: 'Sign up',
+    },
+    hero: {
+      badge: '100% open source · free forever',
+      h1a: 'Your money,',
+      h1b: 'under control.',
+      sub: 'Track expenses, income and investments in one place. Import bank statements, auto-categorize, and understand where every dollar goes.',
+      cta: 'Start free',
+      cta2: 'See demo',
+      trust: 'No credit card · Encrypted data · Open source',
+    },
+    problems: {
+      kicker: 'The problem',
+      title: 'Ever reached month-end with no idea where the money went?',
+      items: [
+        {
+          t: 'Spreadsheets no one updates',
+          d: 'Discipline in week one. Gone by week four. Every single time.',
+        },
+        {
+          t: '"Where did my paycheck go?"',
+          d: 'Credit-card bill arrives and 80% of the charges are a blur.',
+        },
+        {
+          t: 'No full picture',
+          d: 'Investments over there, spending over here, savings somewhere. None of it talks.',
+        },
+      ],
+    },
+    solution: {
+      kicker: 'The solution',
+      title: 'One dashboard. All of your money.',
+      sub: 'Expense Tracker ties accounts, cards and categories into one view that understands both your month and your lifetime totals.',
+    },
+    features: {
+      kicker: 'Features',
+      title: 'Everything you need. Nothing you won\u2019t use.',
+      items: [
+        {
+          tag: '01',
+          t: 'Smart dashboard',
+          d: 'Available balance and "Savings" reflect real accumulated value, while income and expenses respect the monthly filter.',
+        },
+        {
+          tag: '02',
+          t: 'Bank import (CSV)',
+          d: 'Import wizard with auto-detected categories and smart column mapping for your statement.',
+        },
+        {
+          tag: '03',
+          t: 'Custom categories',
+          d: 'Built-in categories plus your own — with emoji and color. Shape it around your life.',
+        },
+        {
+          tag: '04',
+          t: 'Native dark mode',
+          d: 'Dark theme for closing the books late at night. One click to switch.',
+        },
+        {
+          tag: '05',
+          t: 'Supabase security',
+          d: 'Full auth (OAuth, recovery), encrypted database and row-level security on every table.',
+        },
+      ],
+    },
+    how: {
+      kicker: 'How it works',
+      title: 'From sign-up to your first report in under 5 minutes.',
+      steps: [
+        {
+          n: '01',
+          t: 'Create your account',
+          d: 'Email + password or Google login. No paperwork, no card.',
+        },
+        {
+          n: '02',
+          t: 'Import your transactions',
+          d: 'Drop in your bank CSV. We detect categories automatically.',
+        },
+        {
+          n: '03',
+          t: 'See your reports',
+          d: 'Income vs. expense, by category, by month. All real-time.',
+        },
+        {
+          n: '04',
+          t: 'Make decisions',
+          d: 'Understand your spending pattern and reshape what matters.',
+        },
+      ],
+    },
+    cta: {
+      title: 'Ready to own your finances?',
+      sub: 'Five minutes to set up. You can close this tab after.',
+      btn: 'Start free',
+      btn2: 'View on GitHub',
+    },
+    footer: {
+      tag: 'Built with ❤️ by Davyd Fontoura',
+      product: 'Product',
+      company: 'Project',
+      resources: 'Resources',
+      links: {
+        product: [
+          ['Features', '#features'],
+          ['How it works', '#how'],
+          ['Demo', '#'],
+          ['Dark mode', '#'],
+        ],
+        company: [
+          ['GitHub', 'https://github.com/davydfontourac/expense-tracker'],
+          ['Issues', 'https://github.com/davydfontourac/expense-tracker/issues'],
+          ['MIT License', '#'],
+        ],
+        resources: [
+          ['Docs', '#'],
+          ['Changelog', '#'],
+          ['Privacy', '#'],
+        ],
+      },
+      rights: '© 2026 Expense Tracker · Open source under MIT license',
+    },
+    mock: {
+      months: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
+      sidebar: {
+        general: 'GENERAL',
+        dashboard: 'Dashboard',
+        transactions: 'Transactions',
+        categories: 'Categories',
+        savings: 'Savings',
+        analysis: 'ANALYSIS',
+        reports: 'Reports',
+        goals: 'Goals',
+        insights: 'Insights',
+        activeFilter: 'Active filter',
+        october: 'OCTOBER 2026',
+      },
+      header: {
+        greeting: 'Good afternoon, davyd.',
+        spent: 'You spent',
+        ofRevenue: 'of your revenue',
+        inApril: 'in April',
+        april2026: 'April 2026',
+        all: 'All',
+        new: 'New',
+      },
+      kpis: {
+        available: 'AVAILABLE',
+        balance: 'Account balance',
+        income: 'INCOME',
+        thisMonth: 'This month',
+        expenses: 'EXPENSES',
+        vsLastMonth: 'vs. last month',
+      },
+      charts: {
+        monthlyTitle: 'Monthly Evolution',
+        monthlySub: 'Income vs. Expenses · 12 months',
+        janDez: 'JAN → DEC',
+        income: 'Income',
+        expenses: 'Expenses',
+        distTitle: 'Distribution',
+        spent: 'SPENT',
+      },
+      dist: {
+        debit: 'Debit',
+        pix: 'Pix',
+        food: 'Food',
+        transport: 'Transport',
+        others: 'Others',
+      },
+    },
+    visuals: {
+      balance: 'BALANCE',
+      accumulated: 'ACCUMULATED',
+      expenses: 'EXPENSES',
+      oct2026: 'OCT/2026',
+      csv: {
+        food: 'Food',
+        transport: 'Transport',
+        subs: 'Subscriptions',
+        imported: 'transactions imported',
+      },
+      categories: [
+        'Food',
+        'Transport',
+        'Housing',
+        'Leisure',
+        'Health',
+        'Education',
+        'Travel',
+        'New',
+      ],
+    },
+  },
+};
