+++ src/utils/auth-translations.ts
@@ -1,48 +1,154 @@
 export const AUTH_TRANSLATIONS: any = {
   'pt-BR': {
+    splash: { subtitle: 'Seu dinheiro, sob controle.' },
+    onboarding: {
+      step1: { title: 'Rastreie tudo em um só lugar.', desc: 'Despesas, receitas e caixinhas. Sem planilhas, sem fricção.' },
+      step2: { title: 'Importe extratos em 2 cliques.', desc: 'Suporte a CSV de qualquer banco. Nós mapeamos as colunas para você.' },
+      step3: { title: 'Insights que fazem sentido.', desc: 'Categorização automática e relatórios visuais para você economizar.' },
+      skip: 'Pular',
+      next: 'Próximo',
+      finish: 'Começar agora'
+    },
     login: {
       email: 'E-mail',
+      emailPlaceholder: 'Digite seu email',
       password: 'Senha',
+      passwordPlaceholder: 'Digite sua senha',
       forgot: 'Esqueceu a senha?',
       button: 'Entrar',
       buttonLoading: 'Entrando...',
       noAccount: 'Ainda não tem conta?',
-      or: 'ou continue com'
+      noAccountLink: 'Criar conta',
+      or: 'ou continue com',
+      title: 'Bem-vindo de volta',
+      subtitle: 'Entre para gerenciar suas finanças de forma inteligente.',
+      create: 'Crie uma agora',
+      remember: 'Lembrar de mim'
     },
     register: {
       name: 'Nome completo',
+      namePlaceholder: 'Como devemos te chamar?',
       email: 'E-mail',
+      emailPlaceholder: 'Digite seu email',
       password: 'Senha',
+      passwordPlaceholder: 'Mínimo 8 caracteres',
       confirm: 'Confirmar senha',
+      confirmPassword: 'Confirmar Senha',
+      confirmPasswordPlaceholder: 'Confirme sua senha',
       button: 'Criar conta',
       buttonLoading: 'Criando...',
       hasAccount: 'Já tem uma conta?',
+      hasAccountLink: 'Entrar',
       signin: 'Entrar',
       terms: 'Concordo com os Termos e Privacidade.',
-      termsText: 'Ao continuar, você concorda com nossos'
+      termsText: 'Ao continuar, você concorda com nossos',
+      termsLink: 'Termos',
+      termsError: 'Você deve aceitar os termos',
+      title: 'Crie sua conta',
+      subtitle: 'Comece a rastrear seus gastos em menos de 1 minuto.',
+    },
+    forgot: {
+      title: 'Esqueceu a senha?',
+      subtitle: 'Digite seu e-mail e mandaremos um link para você redefinir.',
+      button: 'Enviar link',
+      buttonLoading: 'Enviando...',
+      cancel: 'Cancelar',
+      footer: 'Verifique seu inbox · Spam'
+    },
+    success: {
+      title: 'Verifique seu e-mail',
+      subtitle: 'Enviamos um link de confirmação especial para o seu e-mail. Acesse sua caixa de entrada para ativar sua conta e começar sua jornada.',
+      button: 'Ir para o Login',
+      noEmail: 'Não recebeu?',
+      resend: 'Reenviar link'
+    },
+    menu: {
+      home: 'Página Inicial',
+      theme: 'Tema',
+      language: 'Idioma'
+    },
+    mock: {
+      balance: 'Saldo',
+      lunch: 'iFood - Almoço',
+      today: 'Hoje',
+      transactions: 'transações',
+      apr: 'Abr'
     }
   },
   'en': {
+    splash: { subtitle: 'Your money, under control.' },
+    onboarding: {
+      step1: { title: 'Track everything in one place.', desc: 'Expenses, income and goals. No spreadsheets, no friction.' },
+      step2: { title: 'Import statements in 2 clicks.', desc: 'CSV support for any bank. We map the columns for you.' },
+      step3: { title: 'Insights that make sense.', desc: 'Automatic categorization and visual reports to help you save.' },
+      skip: 'Skip',
+      next: 'Next',
+      finish: 'Get started'
+    },
     login: {
       email: 'Email',
+      emailPlaceholder: 'Enter your email',
       password: 'Password',
+      passwordPlaceholder: 'Enter your password',
       forgot: 'Forgot password?',
       button: 'Sign in',
       buttonLoading: 'Signing in...',
       noAccount: "Don't have an account?",
-      or: 'or continue with'
+      noAccountLink: 'Create one',
+      or: 'or continue with',
+      title: 'Welcome back',
+      subtitle: 'Sign in to manage your finances intelligently.',
+      create: 'Create one now',
+      remember: 'Remember me'
     },
     register: {
       name: 'Full name',
+      namePlaceholder: 'How should we call you?',
       email: 'Email',
+      emailPlaceholder: 'Enter your email',
       password: 'Password',
+      passwordPlaceholder: 'Minimum 8 characters',
       confirm: 'Confirm password',
+      confirmPassword: 'Confirm Password',
+      confirmPasswordPlaceholder: 'Confirm your password',
       button: 'Create account',
       buttonLoading: 'Creating...',
       hasAccount: 'Already have an account?',
+      hasAccountLink: 'Sign in',
       signin: 'Sign in',
       terms: 'I agree to the Terms and Privacy.',
-      termsText: 'By continuing, you agree to our'
+      termsText: 'By continuing, you agree to our',
+      termsLink: 'Terms',
+      termsError: 'You must accept the terms',
+      title: 'Create your account',
+      subtitle: 'Start tracking your expenses in less than 1 minute.',
+    },
+    forgot: {
+      title: 'Forgot password?',
+      subtitle: "Enter your email and we'll send you a link to reset.",
+      button: 'Send link',
+      buttonLoading: 'Sending...',
+      cancel: 'Cancel',
+      footer: 'Check your inbox · Spam'
+    },
+    success: {
+      title: 'Check your email',
+      subtitle: "We've sent a special confirmation link to your email. Check your inbox to activate your account and start your journey.",
+      button: 'Go to Login',
+      noEmail: "Didn't receive it?",
+      resend: 'Resend link'
+    },
+    menu: {
+      home: 'Home Page',
+      theme: 'Theme',
+      language: 'Language'
+    },
+    mock: {
+      balance: 'Balance',
+      lunch: 'iFood - Lunch',
+      today: 'Today',
+      transactions: 'transactions',
+      apr: 'Apr'
     }
   }
 };
